feat: PieceDao 클래스 생성 및 코드 분리
